requirements.txt
redis
google-cloud-redis
